Refactor Kinesis stream providers to extend base Provider and Buffer classes
