+++ kinesis/firehose/src/AWSFirehoseBuffer.js
@@ -1,58 +1,17 @@
-const EventEmitter = require('events');
+const BaseStreamBuffer = require('../../src/BaseStreamBuffer');
 
-class AWSFirehoseBuffer extends EventEmitter {
-    constructor(firehose, params) {
-        super();
-
-        this._streamProvider = firehose;
-        this._params = params;
-        this._messagesByStream = {};
-        this._timeout = null;
+class AWSFirehoseBuffer extends BaseStreamBuffer {
+    _composeRecordData(record) {
+        return {
+            Data: JSON.stringify(record)
+        };
     }
 
-    put(data, streamName) {
-        this._initTimeout();
-
-        if (!this._messagesByStream[streamName]) {
-            this._messagesByStream[streamName] = [];
-        }
-
-        this._messagesByStream[streamName].push(data);
-
-        if (this._messagesByStream[streamName].length >= this._params.maxSize) {
-            this._putBatchToStream(streamName);
-        }
-
-        return Promise.resolve(null);
-    }
-
-    _initTimeout() {
-        if (this._timeout === null && this._params.timeout) {
-            this._timeout = setInterval(() => {
-                Object.keys(this._messagesByStream).forEach(streamName => this._putBatchToStream(streamName));
-            }, this._params.timeout);
-        }
-    }
-
-    _putBatchToStream(streamName) {
-        const records = [];
-
-        this._messagesByStream[streamName].forEach(record => {
-            records.push({
-                Data: JSON.stringify(record)
-            });
-        });
-
-        if (records.length) {
-            this._streamProvider.putRecordBatch({
-                DeliveryStreamName: streamName,
-                Records: records
-            }, (err, response) => {
-                this.emit('putBatch', err, response, streamName);
-            });
-
-            this._messagesByStream[streamName] = [];
-        }
+    _batchRequest(records, streamName, callback = () => {}) {
+        return this._provider.putRecordBatch({
+            DeliveryStreamName: streamName,
+            Records: records
+        }, callback);
     }
 }
