+++ plugin/KinesisPluginService.js
@@ -6,27 +6,27 @@ class KinesisPluginService extends PluginService {
         super();
 
         this._kinesisConfig = kinesisConfig;
-        this._streamProvider = null;
+        this._provider = null;
     }
 
     afterStart() {
         super.afterStart();
-        this._streamProvider = kinesis.createProvider(this._kinesisConfig);
+        this._provider = kinesis.createProvider(this._kinesisConfig);
     }
 
     handleCommand(command) {
         super.handleCommand(command);
-        return this._streamProvider.putCommand(command);
+        return this._provider.putCommand(command);
     }
 
     handleNotification(notification) {
         super.handleNotification(notification);
-        return this._streamProvider.putNotification(notification);
+        return this._provider.putNotification(notification);
     }
 
     handleCommandUpdate(command) {
         super.handleCommandUpdate(command);
-        return this._streamProvider.putCommandUpdate(command);
+        return this._provider.putCommandUpdate(command);
     }
 }
