+++ test/kinesis/unit/dataStreams/src/AWSKinesisDataStreamsBuffer.js
